pyproject.toml
[tool.poetry]
name = "magnifico"
version = "0.1.0"
description = ""
authors = ["guiparpinelli"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.66.0"
pydantic = "^1.8.2"
uvicorn = "^0.14.0"
requests = "^2.26.0"
tinydb = "^4.5.2"
pdfplumber = "^0.5.28"
attrs = "^21.2.0"
cattrs = "^1.8.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
black = { version = "*", allow-prereleases = true }
flake8 = "^4.0.1"
pre-commit = "^2.16.0"
ipython = "^7.30.1"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
exclude = ''

[tool.pytest.ini_options]
minversion = 6.0
log_cli = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
